Fixed number format/font sizes issue in FontPickerController.
The predefined font sizes are now stored in a double
array instead a String array. Before the sizeComboBox
is populated with values, all doubles are converted to
String using the default locale specifc number format.

The text sample for preview is now placed inside a
TextArea. This prevents GUI from becoming unusable and
uncontrollable in case of huge font sizes.
When fonts are too large (e.g. 100 px) then the
sample text covered the sizeComboBox and it was
impossible to revert to old setting.
A restart of the application was required.